Remove databases/php-memcache
This port was added as php70-memcache in 2016, then renamed to php-memcache when
PHP flavor was introduced. It uses the unofficial development from websupport.sk
to support PHP 7.x when pecl-memcache only supports PHP 5.x. Since the
unofficial development from websupport.sk is the official PECL repository now,
it's time to retire this port and use pecl-memcache instead.