devel/dbus: Allow setting the session socket dir during build
Previous to this changeset, the value of TMPDIR in the build environment
would be used as the session socket directory for dbus. For any build
configuration in which TMPDIR may be set in a manner local to the build
environment, using a port-specific value - e.g if using a temporary
directory created under WRKDIR, within the context of each individual
port build, perhaps as similar to a pkgsrc build configuration - dbus'
usage of the build-time TMPDIR may result in a runtime error, under at
least such as the following:

  $ ssh -XY otherhost dbus-launch mate-terminal
  dbus-daemon[...]: Failed to start message bus: Failed to bind
    socket "/usr/ports/devel/dbus/work/[sic]/dbus-[sic]": No such file or directory
  EOF in dbus-launch reading address from bus daemon

Under such a system configuration, the build-time TMPDIR may not exist,
or may not be user-writable, as at the time of that shell command's
evaluation on the otherhost machine.

This change introduces a mk-variable DBUS_SOCKDIR, such that will be
initialized to TMPDIR - retaining the original build configuration - by
default. The variable may be otherwise configured in site make.conf.

The value of DBUS_SOCKDIR will be used during dbus configure, as the
--with-session-socket-dir=${DBUS_SOCKDIR} path, thus overriding the
default behavior of using the build-time TMPDIR (or similar, per
conditional evaluation in configure)

Of course, the DBUS_SOCKDIR must be writable by every user using dbus.

FIXME this changeset does not introduce any documentation about
DBUS_SOCKDIR other than in this changeset message.