lang/ruby2[567]: Add a dtrace option for amd64, i386
This retains some of the previous behavior for dtrace
support, such that the dtace option will be used
only when dtrace is installed under /usr/sbin
on the build host - limiting this to amd64 and i386
architectures, moeover only for port builds on a
FreeBSD OPSYS.

This amends the earlier build configuration, in not
building with dtrace support as a default - even when
dtrace is installed under /usr/sbin. In the latter
regard, perhaps this may be analogous to the debug
support, in the build.

When dtrace is not installed under /usr/sbin and the
dtrace option is enabled, a warning will be emitted.
Alternately, the change could be adapted further,
such as to cause the build to fail - when the port
would be configured for dtrace support, on a build
host not evidently providing dtrace support.

This patch has been tested under the "No dtrace"
configuration, locally, for building with
lang/ruby27